Update index.mdx
The rules for the capitalisation of job titles depending on where it's written. You would capitalise if the job title is used in a heading on a resume, however, you shouldn't capitalise the job title when referring to it in the text of a document like in the original text: 'It doesn't matter if you're a bootcamp grad or Principal engineer'

Reference: https://grammar.yourdictionary.com/capitalization/capitalization-of-job-titles.html

Proposed edit: It doesn't matter if you're a bootcamp grad or principal engineer.